Change working directory before loading test file
bash_unit changes the current working directory to the
one of the code under test just before running the tests.

But it was not changing the current working directory
while loading the test file. Hence, if one had static
code in its test file, this code was run in an unknown
working directory.

Know bash_unit changes the current working directory
before loading the test file.