Display overall result in text mode
See https://github.com/pgrange/bash_unit/pull/71/files

We add a notify_suites_succeded and notify_suites_failed so that
this events may be displayed by any formatter.

We use SUCCESS and FAILURE instead of PASS and FAIL just to stick
with the lingo used in the rest of bash_unit.