More robust install
We add more strict bash flags for install script and we also
recommand using pipe instead of process substitution.

With pipe, the process can catch an error when running curl and
fail but there's no such way of catching an error if we do
process substitution.

This can help having more robust CI.